Add files via upload
How many students were admitted each year? What is the distribution of students across different field study? How many students are enrolled in each specialization? What is the average age of students in each field of study? What is the current semester distribution among all students? Are there any correlations between the field of study and the fees paid? Are there any correlations between the fees and the discount offered on the fees?